+++ lib/src/helpers/src/draw_avatar.dart
@@ -29,7 +29,10 @@ part of '../helpers.dart';
 /// The [shadowColor] is the color of the [BoxShadow], by default it is [Colors.black.withOpacity(0.2)].
 /// The [reverse] is the boolean to reverse shadow of the [BoxDecoration], by default it is false.
 /// [context] is the context of the widget.
-
+@Deprecated(
+  'This utility function will be deleted at version 8.0.0 and will be replaced by the widget `ThemedAvatar` - '
+  'Use the widget `ThemedAvatar` instead',
+)
 Widget drawAvatar({
   Avatar? dynamicAvatar,
   String? avatar,
@@ -43,15 +46,140 @@ Widget drawAvatar({
   bool reverse = false,
   required BuildContext context,
 }) {
-  assert(elevation <= 5, 'The elevation must be less than or equal to 5');
-  assert(elevation >= 0, 'The elevation must be greater than or equal to 0');
-  assert(radius >= 0, 'The radius must be greater than or equal to 0');
+  return ThemedAvatar(
+    dynamicAvatar: dynamicAvatar,
+    avatar: avatar,
+    icon: icon,
+    name: name,
+    size: size,
+    radius: radius,
+    color: color,
+    elevation: elevation,
+    shadowColor: shadowColor,
+    reverse: reverse,
+  );
+}
 
-  Color containerColor = color ?? Theme.of(context).primaryColor;
-  Widget content = const SizedBox();
-  double contentSize = size * 0.4;
+class ThemedAvatar extends StatelessWidget {
+  final Avatar? dynamicAvatar;
+  final String? avatar;
+  final IconData? icon;
+  final String? name;
+  final double size;
+  final double radius;
+  final Color? color;
+  final double elevation;
+  final Color? shadowColor;
+  final bool reverse;
 
-  Color shadow = shadowColor ?? Colors.black.withOpacity(0.2);
+  const ThemedAvatar({
+    super.key,
+
+    /// [dynamicAvatar] is the new avatar engine, can be nullable.
+    this.dynamicAvatar,
+
+    /// [avatar] is the avatar of the user. Can be nullable
+    ///
+    /// Prevent submit [avatar] and [icon] if you are using this.
+    /// Nothing will happend but is less code to you :)
+    this.avatar,
+
+    /// [icon] is the icon of the user. Can be nullable
+    ///
+    /// Prevent submit [avatar] and [icon] if you are using this.
+    /// Nothing will happend but is less code to you :)
+    this.icon,
+
+    /// [name] is the name of the user. Can be nullable
+    this.name,
+
+    /// [size] is the size of the avatar.
+    /// By default, it is 30.
+    this.size = 30,
+
+    /// You can control the radius using [radius] property
+    /// By default, it is 30. (The same as the [size])
+    this.radius = 10,
+
+    /// [color] is the color of the avatar. By default, it is [Theme.of(context).primaryColor].
+    /// Only when [dynamicAvatar] is not null, the default value of the color is based on the
+    /// different types of avatars.
+    /// For [AvatarType.emoji], the default color is [Colors.grey.shade900].
+    /// For [AvatarType.icon], the default color is [Theme.of(context).primaryColor].
+    /// For [AvatarType.base64] and [AvatarType.url], the default color is [Colors.transparent].
+    /// For [AvatarType.none], the default color is [Theme.of(context).primaryColor].
+    this.color,
+
+    /// [elevation] is the elevation of the avatar. By default, it is 1.
+    this.elevation = 1,
+
+    /// The [shadowColor] is the color of the [BoxShadow], by default it is [Colors.black.withOpacity(0.2)].
+    this.shadowColor,
+
+    /// The [reverse] is the boolean to reverse shadow of the [BoxDecoration], by default it is false.
+    this.reverse = false,
+  })  : assert(elevation <= 5, 'The elevation must be less than or equal to 5'),
+        assert(elevation >= 0, 'The elevation must be greater than or equal to 0'),
+        assert(radius >= 0, 'The radius must be greater than or equal to 0');
+
+  @override
+  Widget build(BuildContext context) {
+    Color baseColor = color ?? Theme.of(context).primaryColor;
+    Color baseShadow = shadowColor ?? Colors.black.withOpacity(0.2);
+
+    if (dynamicAvatar != null) {
+      return _renderDynamicAvatar(
+        avatar: dynamicAvatar!,
+        context: context,
+        name: name ?? 'N/A',
+        elevation: elevation,
+        shadowColor: baseShadow,
+        reverse: reverse,
+        size: size,
+        radius: radius,
+        color: baseColor,
+      );
+    }
+
+    if (avatar?.isNotEmpty ?? false) {
+      return _renderImage(
+        image: avatar!,
+        context: context,
+        name: name ?? 'N/A',
+        elevation: elevation,
+        shadowColor: baseShadow,
+        reverse: reverse,
+        size: size,
+        radius: radius,
+        color: baseColor,
+      );
+    }
+
+    if (icon != null) {
+      return _renderIcon(
+        icon: icon!,
+        context: context,
+        name: name ?? 'N/A',
+        elevation: elevation,
+        shadowColor: baseShadow,
+        reverse: reverse,
+        size: size,
+        radius: radius,
+        color: baseColor,
+      );
+    }
+
+    return _generateDeafult(
+      context: context,
+      name: name ?? 'N/A',
+      elevation: elevation,
+      shadowColor: baseShadow,
+      reverse: reverse,
+      size: size,
+      radius: radius,
+      color: baseColor,
+    );
+  }
 
   Widget loadingProgressIndicator(BuildContext context, Widget child, ImageChunkEvent? loadingProgress) {
     if (loadingProgress == null) return child;
@@ -72,143 +200,215 @@ Widget drawAvatar({
     );
   }
 
-  if (dynamicAvatar != null) {
-    switch (dynamicAvatar.type) {
-      case AvatarType.emoji:
-        containerColor = Colors.white;
-        content = Center(
-          child: Text(
-            dynamicAvatar.emoji ?? 'NA',
-            style: TextStyle(
-              fontSize: contentSize,
-              color: validateColor(color: containerColor),
-            ),
-          ),
-        );
-        break;
-      case AvatarType.icon:
-        content = Icon(
-          dynamicAvatar.icon ?? Icons.person,
-          color: validateColor(color: containerColor),
-          size: contentSize,
-        );
-        break;
-      case AvatarType.base64:
-        if ((dynamicAvatar.base64 ?? '').isEmpty) {
-          content = Image.network(
-            'https://cdn.layrz.com/resources/layo/layo.png',
-            fit: BoxFit.cover,
-            filterQuality: FilterQuality.medium,
-            loadingBuilder: loadingProgressIndicator,
-          );
-        } else {
-          content = Image.memory(
-            base64Decode(dynamicAvatar.base64!.split(',').last),
-            fit: BoxFit.cover,
-            filterQuality: FilterQuality.medium,
-          );
-        }
-        break;
-      case AvatarType.url:
-        content = Image.network(
-          dynamicAvatar.url ?? 'https://cdn.layrz.com/resources/layo/layo.png',
-          fit: BoxFit.cover,
-          filterQuality: FilterQuality.medium,
-          loadingBuilder: loadingProgressIndicator,
-        );
-        break;
-      default:
-        content = Center(
-          child: Text(
-            name != null
-                ? name.length >= 2
-                    ? name.substring(0, 2).toUpperCase()
-                    : name
-                : 'NA',
-            style: TextStyle(
-              color: validateColor(color: containerColor),
-              fontSize: contentSize,
-            ),
-          ),
-        );
-        break;
-    }
-  } else if (avatar != null && avatar.isNotEmpty) {
-    containerColor = color ?? Colors.white;
-    if (avatar.startsWith('data:')) {
-      content = Image.memory(
-        base64Decode(avatar.split(',').last),
-        fit: BoxFit.cover,
-        filterQuality: FilterQuality.medium,
-      );
-    } else if (avatar.startsWith('http')) {
-      content = Image.network(
-        avatar,
-        fit: BoxFit.cover,
-        filterQuality: FilterQuality.medium,
-        loadingBuilder: loadingProgressIndicator,
-      );
-    } else if (avatar.isEmpty) {
-      content = Image.network(
-        'https://cdn.layrz.com/resources/layo/layo.png',
-        fit: BoxFit.cover,
-        filterQuality: FilterQuality.medium,
-        loadingBuilder: loadingProgressIndicator,
-      );
-    } else {
-      content = Image.asset(
-        avatar,
-        fit: BoxFit.cover,
-        filterQuality: FilterQuality.medium,
+  Widget _renderImage({
+    required String image,
+    required BuildContext context,
+    required String name,
+    required double elevation,
+    required Color shadowColor,
+    required bool reverse,
+    required double size,
+    required double radius,
+    required Color color,
+  }) {
+    if (image.isEmpty) {
+      return _generateDeafult(
+        context: context,
+        elevation: elevation,
+        shadowColor: shadowColor,
+        reverse: reverse,
+        color: color,
+        size: size,
+        radius: radius,
+        name: name,
       );
     }
-  } else if (icon != null) {
-    content = Center(
-      child: Icon(
-        icon,
-        color: validateColor(color: containerColor),
-        size: contentSize,
+
+    return _generateCircle(
+      context: context,
+      elevation: elevation,
+      shadowColor: shadowColor,
+      reverse: reverse,
+      color: color,
+      size: size,
+      radius: radius,
+      child: ThemedImage(
+        path: image,
+        width: size,
+        height: size,
+        fit: BoxFit.cover,
       ),
     );
-  } else {
-    content = Center(
-      child: Text(
-        name != null
-            ? name.length >= 2
-                ? name.substring(0, 2).toUpperCase()
-                : name
-            : 'NA',
-        style: TextStyle(
-          color: validateColor(color: containerColor),
-          fontSize: contentSize,
+  }
+
+  Widget _renderIcon({
+    required IconData icon,
+    required BuildContext context,
+    required String name,
+    required double elevation,
+    required Color shadowColor,
+    required bool reverse,
+    required double size,
+    required double radius,
+    required Color color,
+  }) {
+    return _generateCircle(
+      context: context,
+      elevation: elevation,
+      shadowColor: shadowColor,
+      reverse: reverse,
+      color: color,
+      size: size,
+      radius: radius,
+      child: Icon(
+        icon,
+        color: validateColor(color: color),
+        size: size * 0.7,
+      ),
+    );
+  }
+
+  Widget _renderDynamicAvatar({
+    required Avatar avatar,
+    required BuildContext context,
+    required String name,
+    required double elevation,
+    required Color shadowColor,
+    required bool reverse,
+    required double size,
+    required double radius,
+    required Color color,
+  }) {
+    switch (avatar.type) {
+      case AvatarType.emoji:
+        return _generateCircle(
+          context: context,
+          elevation: elevation,
+          shadowColor: shadowColor,
+          reverse: reverse,
+          color: Colors.white,
+          size: size,
+          radius: radius,
+          child: Text(
+            avatar.emoji ?? _cleanName(name),
+            style: Theme.of(context).textTheme.bodyLarge?.copyWith(fontSize: size * 0.6),
+          ),
+        );
+      case AvatarType.icon:
+        return _renderIcon(
+          icon: avatar.icon ?? Icons.person,
+          context: context,
+          name: name,
+          elevation: elevation,
+          shadowColor: shadowColor,
+          reverse: reverse,
+          size: size,
+          radius: radius,
+          color: color,
+        );
+
+      case AvatarType.base64:
+        return _renderImage(
+          image: avatar.base64 ?? '',
+          context: context,
+          name: name,
+          elevation: elevation,
+          shadowColor: shadowColor,
+          reverse: reverse,
+          size: size,
+          radius: radius,
+          color: color,
+        );
+
+      case AvatarType.url:
+        return _renderImage(
+          image: avatar.url ?? '',
+          context: context,
+          name: name,
+          elevation: elevation,
+          shadowColor: shadowColor,
+          reverse: reverse,
+          size: size,
+          radius: radius,
+          color: color,
+        );
+
+      default:
+        return _generateDeafult(
+          context: context,
+          elevation: elevation,
+          shadowColor: shadowColor,
+          reverse: reverse,
+          color: color,
+          size: size,
+          radius: radius,
+          name: name,
+        );
+    }
+  }
+
+  Widget _generateDeafult({
+    required BuildContext context,
+    required String name,
+    required double elevation,
+    required Color shadowColor,
+    required bool reverse,
+    required double size,
+    required double radius,
+    required Color color,
+  }) {
+    return _generateCircle(
+      context: context,
+      elevation: elevation,
+      shadowColor: shadowColor,
+      reverse: reverse,
+      color: color,
+      size: size,
+      radius: radius,
+      child: Center(
+        child: Text(
+          _cleanName(name),
+          style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                color: validateColor(color: color),
+                fontSize: size * 0.4,
+              ),
         ),
       ),
     );
   }
 
-  return Container(
-    width: size,
-    height: size,
-    clipBehavior: Clip.antiAlias,
-    decoration: BoxDecoration(
-      color: containerColor,
-      borderRadius: BorderRadius.circular(radius),
-      border: elevation == 0
-          ? Border.all(
-              color: shadow,
-              width: 1,
-            )
-          : null,
-      boxShadow: elevation > 0
-          ? [
-              BoxShadow(
-                color: shadow,
-                blurRadius: 2 * elevation.toDouble(),
-                offset: Offset(0, elevation.toDouble() * (reverse ? -1 : 1)), // changes position of shadow
-              ),
-            ]
-          : null,
-    ),
-    child: content,
-  );
+  Widget _generateCircle({
+    required Widget child,
+    required BuildContext context,
+    required double elevation,
+    required Color shadowColor,
+    required bool reverse,
+    required double size,
+    required double radius,
+    required Color color,
+  }) {
+    return Container(
+      width: size,
+      height: size,
+      decoration: generateContainerElevation(
+        context: context,
+        elevation: elevation,
+        color: color,
+        shadowColor: shadowColor,
+        reverse: reverse,
+        radius: radius,
+      ),
+      alignment: Alignment.center,
+      clipBehavior: Clip.antiAlias,
+      child: child,
+    );
+  }
+
+  String _cleanName(String? raw) {
+    if (raw == null) return 'NA';
+    String output = raw.replaceAll(RegExp(r'[^a-zA-Z0-9]'), '');
+    if (output.isEmpty) return 'NA';
+    if (output.length < 2) return output.toLowerCase();
+    return output.substring(0, 2).toUpperCase();
+  }
 }